backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Ethan Brown <ethan@ewbrowntech.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
sqlalchemy = "^2.0.25"
httpx = "^0.26.0"
asyncpg = "^0.29.0"
python-multipart = "^0.0.6"
aiosqlite = "^0.19.0"
alembic = "^1.13.1"
python-magic = "^0.4.27"
apscheduler = "^3.10.4"
slowapi = "^0.1.8"
passlib = "^1.7.4"
bcrypt = "4.0.1"
pyjwt = "^2.8.0"
python-jose = "^3.3.0"


[tool.poetry.group.dev.dependencies]
black = "^23.12.1"
pytest-cov = "^4.1.0"
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
